backend/requirements.txt
alembic==1.15.2
bcrypt==4.2.1
httpx==0.28.1
celery[redis]==5.4.0
email-validator==2.2.0
fastapi==0.115.12
passlib[bcrypt]==1.7.4
pydantic==2.11.1
pydantic-settings==2.8.1
psycopg[binary]==3.2.6
pytest==8.3.5
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
sqlalchemy==2.0.39
uvicorn[standard]==0.34.0
